+++ core/docs/changelog.md
@@ -1,9 +1,10 @@
 # TBD
 ### Changes
 * The API container now assumes the enclave data volume is a directory on the Docker host machine, and bind-mounts it to the containers it starts rather than via volume-mounts
-
-### Changes
 * Swapped the overly-complex `V0LaunchArgs` back to the old way, of a simple `APIContainerLauncher`
+
+### Breaking Changes
+* Upgraded to Kurt Client vTODOOOOOO, which renames several object properties
 
 # 1.26.4
 ### Changes